requirements.txt
gamspy
folium
pyyaml
requests
graphviz
bokeh